Phase 8: contract deadline failures, MIA respawn tests, retired kerbals UI
Task 32 — Contract deadline failure generation:
- Capture deadline + failure penalties at contract accept time (GameStateRecorder)
- Parse structured detail format with backward compat for legacy recordings (GameStateEventConverter)
- Serialize fundsPenalty/repPenalty on ContractAccept actions (GameAction)
- Inject synthetic ContractFail actions in ContractsModule.PrePass for expired deadlines
- Re-sort action list after PrePass in RecalculationEngine
- 12 new tests (8 ContractsModule, 4 GameStateEventConverter)

Task 36 — MIA respawn override verified + tested:
- Confirmed ApplyToRoster Step 3 already overrides KSP respawns
- 4 new MIA respawn override tests in KerbalReservationTests
- Updated D7 status to Done in deferred items doc

Task 37 — Retired Stand-ins UI section:
- Show retired kerbals list in Game Actions window when non-empty
- Change-based logging (not per-frame)

Also: fix Reputation.Instance namespace collision (global::), fix Assert.Equal
ambiguity in pre-existing tests, update T42 reference in LedgerOrchestrator

+++ Source/Parsek.Tests/LedgerOrchestratorTests.cs
@@ -254,7 +254,7 @@ namespace Parsek.Tests
             Assert.Single(actions);
             Assert.Equal(GameActionType.FundsSpending, actions[0].Type);
             Assert.Equal(FundsSpendingSource.VesselBuild, actions[0].FundsSpendingSource);
-            Assert.Equal(10000f, actions[0].FundsSpent, 1);
+            Assert.Equal(10000.0, (double)actions[0].FundsSpent, 1);
             Assert.Equal(100.0, actions[0].UT);
             Assert.Equal("rec-build-cost", actions[0].RecordingId);
 
@@ -287,12 +287,12 @@ namespace Parsek.Tests
             var buildAction = actions[0];
             Assert.Equal(GameActionType.FundsSpending, buildAction.Type);
             Assert.Equal(FundsSpendingSource.VesselBuild, buildAction.FundsSpendingSource);
-            Assert.Equal(10000f, buildAction.FundsSpent, 1);
+            Assert.Equal(10000.0, (double)buildAction.FundsSpent, 1);
 
             var recoveryAction = actions[1];
             Assert.Equal(GameActionType.FundsEarning, recoveryAction.Type);
             Assert.Equal(FundsEarningSource.Recovery, recoveryAction.FundsSource);
-            Assert.Equal(7000f, recoveryAction.FundsAwarded, 1);
+            Assert.Equal(7000.0, (double)recoveryAction.FundsAwarded, 1);
             Assert.Equal(300.0, recoveryAction.UT);
             Assert.Equal("rec-recovery", recoveryAction.RecordingId);
 
@@ -397,7 +397,7 @@ namespace Parsek.Tests
             // Build cost only (50000-47000=3000), no recovery (need 2+ points)
             Assert.Single(actions);
             Assert.Equal(GameActionType.FundsSpending, actions[0].Type);
-            Assert.Equal(3000f, actions[0].FundsSpent, 1);
+            Assert.Equal(3000.0, (double)actions[0].FundsSpent, 1);
 
             RecordingStore.ResetForTesting();
         }
